requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
asgiref=3.5.2=py310hecd8cb5_0
bzip2=1.0.8=h1de35cc_0
ca-certificates=2022.10.11=hecd8cb5_0
certifi=2022.9.24=py310hecd8cb5_0
django=4.1=py310hecd8cb5_0
django-filter=22.1=pypi_0
djangorestframework=3.14.0=pypi_0
djangorestframework-simplejwt=5.2.2=pypi_0
libcxx=14.0.6=h9765a3e_0
libffi=3.3=hb1e8313_2
markdown=3.4.1=pypi_0
ncurses=6.3=hca72f7f_3
openssl=1.1.1s=hca72f7f_0
pip=22.2.2=py310hecd8cb5_0
pyjwt=2.6.0=pypi_0
python=3.10.6=hdfd78df_1
pytz=2022.6=pypi_0
readline=8.2=hca72f7f_0
setuptools=65.5.0=py310hecd8cb5_0
sqlite=3.39.3=h707629a_0
sqlparse=0.4.1=py_0
tk=8.6.12=h5d9f67b_0
tzdata=2022f=h04d1e81_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.6=hca72f7f_0
zlib=1.2.13=h4dc903c_0
